remove ts-node and add interactive mode
interactive mode is reading a spec file from stdin.

If I paste a list of spec like that:
```
../../dist/addition8.spec.js
../../dist/addition21.spec.js
../../dist/addition31.spec.js
../../dist/addition98.spec.js
../../dist/addition88.spec.js
../../dist/addition43.spec.js
../../dist/addition53.spec.js
../../dist/addition65.spec.js
../../dist/addition75.spec.js
```
with 100 tests, it is executing them in 81ms, quite fast. Typescript
version is way slower: 3.8s
This is probably ts-node being slow, I wonder if it would not be way
more efficient, to just do tsc and then work with the compiled js
files.

And Jest is taking 700ms with multiple workers, and almost 2s with
runInBand.