initialize go module for main process
And just create a basic structure which is doing:
- file walk to find the specs
- putting all specs found in a chan
- listener of this chan is dispatching the found specs to workers

This is taking around 200ms to find all spec files (when measuring
with `time`), which is not so great, it was taking 40ms before the
introduction of workers.

Wondering if it would be better to directly use a chan with a big
size, like this it will act as the buffer instead of having this
indirection with specChan and workerChan and buffer.