simplify main process
No intermediate goroutine to dispatch to workers and buffer the
specs (see previous commit message). It doesn't seems to impact
performance to just have a chan with a buffer being equals to the
number of workers.

Some sleep has been added to the worker (10ms), shutdown has then
been fixed as I didn't realized I was cancelling the context too
early, we need to ensure everything has been consumed before.

And this is nice to run the program with 1 worker, then 2, 3, ...
and see the time decreasing accordingly.

```
$ time ./yatr 1 sample/100specs
./yatr 1 sample/100specs  0.01s user 0.02s system 2% cpu 1.099 total

$ time ./yatr 10 sample/100specs
./yatr 10 sample/100specs  0.01s user 0.03s system 27% cpu 0.124 total
```